Started fixing question

+++ src/loops/simple_programs_with_repitition/medium/q4/SerieslyQuestionMark.java
@@ -11,9 +11,9 @@ public class SerieslyQuestionMark {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter a number for the upper bound of the series: ");
-        int n = input.nextInt();
+        double n = input.nextInt();
         double sum = 0;
-        for (double i = 1; i < n; i++)
+        for (int i = 1; i < n; i++)
             sum += i / n;
         System.out.println("The sum of the series is " + sum);
     }
